+++ neurom/viewer.py
@@ -38,13 +38,16 @@ Examples:
     >>> viewer.draw(nrn, mode='dendrogram') # dendrogram plot
 """
 
-from neurom.view.view import (plot_neuron, plot_neuron3d,
-                              plot_tree, plot_tree3d,
-                              plot_soma, plot_soma3d,
-                              plot_dendrogram)
-from neurom.view import common
+from neurom.view.matplotlib_impl import (plot_neuron, plot_neuron3d,
+                                         plot_tree, plot_tree3d,
+                                         plot_soma, plot_soma3d,
+                                         plot_dendrogram)
+from neurom.view import matplotlib_utils
 from neurom.core.neuron import Section, Neurite, Neuron
 from neurom.core.soma import Soma
+from neurom.utils import deprecated_module
+
+deprecated_module('Module `viewer` is deprecated. See the documentation\'s migration page.')
 
 MODES = ('2d', '3d', 'dendrogram')
 
@@ -86,7 +89,8 @@ def draw(obj, mode='2d', **kwargs):
         NotDrawableError if obj type and mode combination is not drawable
 
     Examples:
-        >>> nrn = ... # load a neuron
+        >>> from neurom import viewer, load_neuron
+        >>> nrn = load_neuron('/path/to/morphology') # load a neuron
         >>> fig, _ = viewer.draw(nrn)             # 2d plot
         >>> fig.show()
         >>> fig3d, _ = viewer.draw(nrn, mode='3d') # 3d plot
@@ -102,8 +106,8 @@ def draw(obj, mode='2d', **kwargs):
             raise NotImplementedError('Option realistic_diameter not implemented for 3D plots')
         del kwargs['realistic_diameters']
 
-    fig, ax = (common.get_figure() if mode in ('2d', 'dendrogram')
-               else common.get_figure(params={'projection': '3d'}))
+    fig, ax = (matplotlib_utils.get_figure() if mode in ('2d', 'dendrogram')
+               else matplotlib_utils.get_figure(params={'projection': '3d'}))
 
     if isinstance(obj, Neuron):
         tag = 'neuron'
@@ -121,12 +125,12 @@ def draw(obj, mode='2d', **kwargs):
         raise NotDrawableError('No drawer for class %s, mode=%s' % (obj.__class__, mode)) from e
 
     output_path = kwargs.pop('output_path', None)
-    plotter(ax, obj, **kwargs)
+    plotter(obj, ax, **kwargs)
 
     if mode != 'dendrogram':
-        common.plot_style(fig=fig, ax=ax, **kwargs)
+        matplotlib_utils.plot_style(fig=fig, ax=ax, **kwargs)
 
     if output_path:
-        common.save_plot(fig=fig, output_path=output_path, **kwargs)
+        matplotlib_utils.save_plot(fig=fig, output_path=output_path, **kwargs)
 
     return fig, ax